requirements.txt
# ============================================================
# requirements.txt  (RUNTIME)
#
# Inference + API only.
# Must work in python:slim Docker WITHOUT system GUI libs.
#
# Key fix:
# - Use opencv-python-headless instead of opencv-python
#   to avoid libGL.so.1 dependency.
# ============================================================


# -----------------------------
# General data / utils
# -----------------------------
numpy<2.0.0
pandas==2.2.3


# -----------------------------
# Google Cloud
# (optional model download from GCS)
# -----------------------------
google-cloud-storage>=2.16.0


# -----------------------------
# API
# -----------------------------
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9   # for UploadFile endpoints


# -----------------------------
# Imaging
# -----------------------------
Pillow>=10.2.0

# ✅ Headless OpenCV (NO libGL needed in Docker slim)
opencv-python-headless>=4.9.0.80


# -----------------------------
# Torch / vision stack
# -----------------------------
torch>=2.1.0
torchvision>=0.16.0
torchxrayvision>=1.2.1


# -----------------------------
# BLIP inference (Transformers)
# -----------------------------
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.2


# -----------------------------
# Hugging Face datasets tools
# (optional but safe)
# -----------------------------
datasets>=2.19.0


# -----------------------------
# Progress / logging
# -----------------------------
tqdm>=4.66.0


# -----------------------------
# Configuration
# -----------------------------
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
